fix the problem sprite can't  be location
in the beginning  my getSpriteFrameByName method can’t get the right
frame, at first I though there were something wrong in this
getSpriteFrameByName method so I try to fix it, I test this method in
helloworldScene by change structure of this method but it prove a waste
of time, then I try to create a Sprite by using creat(filename,rect)
and it failed again, everytime I change the rect’s value, it comes to
change position out of my thought, I could not control the frame’s
range by using rect, and the setpostion method also useless. After all
of my work, I found out the problem, in the AppDelegate there is a
function set the display range, and the default value is more than 1024
and it can change with the device, so I change this value to 288, 512
and all the problem goes, thank gods! this problem take me 3 days